fix print_version() to return where it came from
The print_version() is jumping through hoops to get back to where
it was called from, though it actually does get called with "bl"
already, so it knows.

Just return to where we came from after print_version(), making it
a function we can call from other places in the code.

Signed-off-by: Alexander Graf <agraf@suse.de>