Work around missing sc 1 traps on pHyp
When running a pseries guest in PR KVM on top of pHyp, sc 1 instructions
are handled directly by pHyp, so we don't get to see them.

That means we need to get inventive. Invent a new instruction that behaves
like sc 1, but really is a reserved instruction that traps. This instruction
can be used by KVM to emulate sc 1 behavior.

This patch adds the SLOF support for it. With this, SLOF detects whether
it's running on such a broken setup and if so patches itself to execute
the fake sc 1 instruction instead of the real one.

Furthermore, we also hook into "quiesce" which Linux calls when it boots.
This gives us the chance to also patch Linux when it boots up, so it uses
the fake sc 1 too.

Signed-off-by: Alexander Graf <agraf@suse.de>

---

v1 -> v2:

  - create TOC instead of pointer to pointer to function

v2 -> v3:

  - rebase on current git
  - fix js2x build

v3 -> v4:

  - move check function into libhvcall
  - restrict patch scope to SLOF's loader's .text, paflof's .text, RTAS
    and [ MAX(r2-32MB, 0) ... r2 ] when quiesce gets called

v4 -> v5:

  - rename hv-patch-sc1 to check-and-patch-sc1
  - rename call-hv-patch-sc1 to check-patch-kernel-sc1
  - use kernel return pointer to know where the kernel is located
  - add little endian check to swizzle byte reversed instructions
  - add sc1 checking cache to prevent us from clobbering DABR
  - rework sc1 patching code to be more readable

v5 -> v6:

  - fix compile warnings / breakages
  - add debug prints
  - trim search range to .text section for Linux kernels